Update README.md
This is the addition of my .R file.

I could not, after an hour of trying, get the file committed.

I'm going to take a shower now and cry (I understand this may be an inappropriate comment and completely unprofessional and for that, I apologize).  
I completed the first two classes and got certificates with distinction.
For this class, I now run the risk of not even earning my certificate.

If you read this much, I again apologize for my "whininess".